Fixed filestructure
